Move requirement lowering into its own module
